+++ Editor/Scripts/EditorWindows/InputWrapperDebuggerWindow.cs
@@ -1,4 +1,7 @@
+using System;
+using System.Collections.Generic;
 using NPTP.InputSystemWrapper.Data;
+using NPTP.InputSystemWrapper.Enums;
 using UnityEditor;
 using UnityEngine;
 using FontStyle = UnityEngine.FontStyle;
@@ -8,6 +11,21 @@ namespace NPTP.InputSystemWrapper.Editor.EditorWindows
 	public class InputWrapperDebuggerWindow : EditorWindow
 	{
 		private const string EMPTY = "";
+		private const int MAX_SHOWN_RECENT_CONTEXTS = 3;
+
+		private class TimestampedObject<T>
+		{
+			public T Value { get; }
+			public string Timestamp { get; }
+
+			public TimestampedObject(T value, string timestamp)
+			{
+				Value = value;
+				Timestamp = timestamp;
+			}
+		}
+		
+		private List<TimestampedObject<InputContext>> mostRecentContexts = new();
 
 		private OfflineInputData offlineInputData;
 		private OfflineInputData OfflineInputData
@@ -16,6 +34,43 @@ namespace NPTP.InputSystemWrapper.Editor.EditorWindows
 			{
 				if (offlineInputData == null) offlineInputData = Helper.OfflineInputData;
 				return offlineInputData;
+			}
+		}
+
+		private void OnEnable()
+		{
+			EditorApplication.playModeStateChanged -= HandlePlayModeStateChanged;
+			EditorApplication.playModeStateChanged += HandlePlayModeStateChanged;
+		}
+
+		private void OnDisable()
+		{
+			EditorApplication.playModeStateChanged -= HandlePlayModeStateChanged;
+		}
+
+		private void HandlePlayModeStateChanged(PlayModeStateChange state)
+		{
+			
+			switch (state)
+			{
+				case PlayModeStateChange.EnteredPlayMode:
+					mostRecentContexts.Clear();
+					mostRecentContexts.Add(new TimestampedObject<InputContext>(Input.Context, 0.ToString()));
+					Input.EDITOR_OnPlayerInputContextChanged += HandlePlayerInputContextChanged;
+					break;
+				case PlayModeStateChange.ExitingPlayMode:
+					Input.EDITOR_OnPlayerInputContextChanged -= HandlePlayerInputContextChanged;
+					break;
+			}
+		}
+
+		private void HandlePlayerInputContextChanged(PlayerID playerID, InputContext inputContext)
+		{
+			NPTPDebug.Log($"Input Context changed for {playerID}: {inputContext}");
+			mostRecentContexts.Add(new TimestampedObject<InputContext>(inputContext, Time.frameCount.ToString()));
+			if (mostRecentContexts.Count > MAX_SHOWN_RECENT_CONTEXTS)
+			{
+				mostRecentContexts.RemoveAt(0);
 			}
 		}
 
@@ -42,15 +97,23 @@ namespace NPTP.InputSystemWrapper.Editor.EditorWindows
 			GUILayout.BeginVertical();
 			ShowDebugInfoField("Current Control Scheme", Input.CurrentControlScheme.ToString());
 			ShowDebugInfoField("Current Context", Input.Context.ToString());
-			ShowActiveMaps();
+			ShowIndentedField("Active Maps", ActiveMapLabelFields);
+			ShowIndentedField("Most Recent Contexts", MostRecentContextLabelFields);
 			GUILayout.EndVertical();
 		}
 
-		private void ShowActiveMaps()
+		private void ShowIndentedField(string fieldName, Action showAction)
 		{
-			ShowDebugInfoField("Active Maps");
+			ShowDebugInfoField(fieldName);
 			EditorGUI.indentLevel++;
 			EditorGUILayout.BeginVertical();
+			showAction?.Invoke();
+			EditorGUILayout.EndVertical();
+			EditorGUI.indentLevel--;
+		}
+
+		private void ActiveMapLabelFields()
+		{
 			foreach (InputContextInfo inputContextInfo in OfflineInputData.InputContexts)
 			{
 				if (inputContextInfo.Name.AsEnumMember() != Input.Context.ToString())
@@ -65,9 +128,14 @@ namespace NPTP.InputSystemWrapper.Editor.EditorWindows
 
 				break;
 			}
+		}
 
-			EditorGUILayout.EndVertical();
-			EditorGUI.indentLevel--;
+		private void MostRecentContextLabelFields()
+		{
+			for (int i = mostRecentContexts.Count - 1; i >= 0; i--)
+			{
+				EditorGUILayout.LabelField($"{mostRecentContexts[i].Value} [Frame {mostRecentContexts[i].Timestamp}]");
+			}
 		}
 
 		private static void ShowDebugInfoField(string boldLabel, string info = EMPTY)