+++ Runtime/Scripts/Data/InputContextInfo.cs
@@ -9,13 +9,15 @@ namespace NPTP.InputSystemWrapper.Data
     {
         [SerializeField] private string name;
         public string Name => name;
-
-        [InputMapSelector]
-        [SerializeField] private string[] activeMaps;
-        public string[] ActiveMaps => activeMaps;
-
+        
         [SerializeField] private bool enableKeyboardTextInput;
         public bool EnableKeyboardTextInput => enableKeyboardTextInput;
+
+        [InputMapSelector][SerializeField] private string[] activeMaps;
+        public string[] ActiveMaps => activeMaps;
+        
+        [SerializeField] private EventSystemActionSpecification[] eventSystemActionOverrides;
+        public EventSystemActionSpecification[] EventSystemActionOverrides => eventSystemActionOverrides;
 
 #if UNITY_EDITOR
         public void EDITOR_SetName(string n)