+++ Runtime/Scripts/Enums/EventSystemActionType.cs
@@ -1,0 +1,21 @@
+namespace NPTP.InputSystemWrapper.Enums
+{
+    public enum EventSystemActionType
+    {
+        Point,
+        LeftClick,
+        MiddleClick,
+        RightClick,
+        ScrollWheel,
+        Move,
+        Submit,
+        Cancel,
+        TrackedDevicePosition,
+        TrackedDeviceOrientation        
+    }
+
+    internal static class EventSystemActionTypeExtensions
+    {
+        
+    }
+}